LS Multiplet Oscillator Strength Calculator 0.1
New updated version of the LS Multiplet code, makes direct use of the formulas I found in Cowan 1981 to calculate log(gf) values for both fine and hyperfine structure lines.

Requires a linelist for automatic detection of secondary lines, recommended to use the LineListReader.py I've written. Given a wavelength and an astrophysical log(gf) value the code will identify which line is being investigated, search for secondary lines, identifying if they are fine or hyperfine structure, and return log(gf) values scaled to the given value for the secondary lines. The hyperfine structure code is a bit janky as the VALD3 linelist does not include the F numbers for the transitions. Instead all possible hyperfine transitions are calculated, and matched to existing lines by assunming that their relative strengths should remain the same.