Weak Line Approximation strength 0.1
Calculates a relative strength for a spectral line based on the weak-line approximation. Note that as only log(gf), excitation energy and temperature are taken into account, comparison of strength values are only applicable for lines of the same element and ionisation stage.
Useful for figuring out which lines are interesting and which are blends.